+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add a delete property button toggle for frontmatter fields in the Form view.
+
 ## [0.5.0] - 2026-08-09
 
 ### Added